% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.onLoad}
\alias{.onLoad}
\title{Connect to the Lahman database and create
accessor funtions when the package is loaded.}
\usage{
.onLoad(libname, pkgname)
}
\description{
Note how the exportPattern below matches the con_id
("lahman") in the call to dbc_init. This will ensure
all of your table accessor functions are exported to
the package namespace.
}
