% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connections.R
\name{get_lahman_connection}
\alias{get_lahman_connection}
\title{Retrieve a connection from to the Lahman database.}
\usage{
get_lahman_connection()
}
\description{
This example is simple, but you will most likely use
DBI::dbConnect for your connection. If connecting to
a secure database, you can store your credentials in
your .Renviron and utilize them for this function.
}
